Refactor Sinkhorn Bures test
